refactor(arch): move sync types from core.sync to teatree.types
SyncResult, PREntry, DiscussionSummary, SyncBackend, RawAPIDict,
PREntryDict, and cache key constants now live in teatree.types — a
leaf module with no Django or teatree.core dependencies. This reduces
the backends → core coupling: backends that only need type definitions
no longer pull in the full core.sync module.

core/sync.py retains the orchestration functions (sync_followup,
_merge_results, _overlay_name, fetch_notion_statuses) and imports
only SyncBackend + SyncResult from types.

28 files updated (22 src + 6 tests).
