+++ src/teatree/types.py
@@ -4,8 +4,9 @@ These types have no Django dependencies and no imports from ``teatree.core``,
 so they can be used by any layer without introducing cycles.
 """
 
+from abc import ABC, abstractmethod
 from collections.abc import Callable
-from dataclasses import dataclass, field
+from dataclasses import asdict, dataclass, field
 from pathlib import Path
 from typing import TypedDict
 
@@ -109,3 +110,82 @@ class ProvisionStep:
     callable: Callable[[], None]
     required: bool = True
     description: str = ""
+
+
+# ── Sync types (shared vocabulary between core and backends) ─────────
+
+LAST_SYNC_CACHE_KEY = "teatree_followup_last_sync"
+PENDING_REVIEWS_CACHE_KEY = "teatree_pending_reviews"
+
+type RawAPIDict = dict[str, object]
+type PREntryDict = dict[str, object]
+
+
+@dataclass(slots=True)
+class SyncResult:
+    prs_found: int = 0
+    issues_found: int = 0
+    tickets_created: int = 0
+    tickets_updated: int = 0
+    labels_fetched: int = 0
+    prs_merged: int = 0
+    prs_closed: int = 0
+    reviews_synced: int = 0
+    worktrees_cleaned: int = 0
+    errors: list[str] = field(default_factory=list)
+
+
+@dataclass(frozen=True, slots=True)
+class DiscussionSummary:
+    status: str
+    detail: str
+
+    def to_dict(self) -> dict[str, str]:
+        return asdict(self)
+
+
+@dataclass(slots=True)
+class PREntry:
+    url: str
+    title: str
+    branch: str
+    draft: bool
+    repo: str
+    iid: int
+    updated_at: str
+    state: str = "opened"
+    pipeline_status: str | None = None
+    pipeline_url: str | None = None
+    approvals: RawAPIDict | None = None
+    discussions: list[DiscussionSummary] | None = None
+    e2e_test_plan_url: str | None = None
+    review_requested: bool | None = None
+    reviewer_names: list[str] | None = None
+    review_permalink: str | None = None
+    review_channel: str | None = None
+    notion_status: str | None = None
+    notion_url: str | None = None
+    draft_comments_pending: bool | None = None
+    draft_comments_count: int | None = None
+    approvals_dismissed_at: str | None = None
+    dismissed_approvers: list[str] | None = None
+
+    def to_dict(self) -> PREntryDict:
+        result: PREntryDict = {}
+        for k in self.__slots__:
+            v = getattr(self, k)
+            if v is None:
+                continue
+            if k == "discussions":
+                result[k] = [d.to_dict() for d in v]
+            else:
+                result[k] = v
+        return result
+
+
+class SyncBackend(ABC):
+    @abstractmethod
+    def is_configured(self, overlay: object) -> bool: ...
+
+    @abstractmethod
+    def sync(self, overlay: object) -> SyncResult: ...
